Remove unnecessary CPU ByteBuffer allocation
Previously a memory allocation was done for glTexImage2D. This can
be done on the GPU automatically by passing in NULL.

No obvious performance improvement was noticed by removing this
allocation but this should allow the allocation to happen within
the GPU optimally.

Alternatively, a ring-buffer could be created to manage a CPU
allocations but this will be more complex and likely slower than
allowing the GPU to handle this.