DO NOT MERGE: specific test branch for exo bug
A test branch to specifically demostrate a bug on Chrome OS when
exoplayer surfaces are disconnected and reconnected.

The use-case is if the decode surfaces are Views which are reconstructed
during rotation events. After rotation when they are reconnected to the
Player there can be an error. Looks like a timing issue. Reproduces
often on Chrome OS and rarely on Android.