Reduce circular buffer de/allocations
With large video files, the number of allocations and deallocations for
audio buffers can be large. Even though memory use remains reasonable
and the buffers are small, the sheer number of de/allocations can gum
things up. This reduces copies and uses circular buffers to reduce
allocations as the pipeline proceeds.